Add kafka client metrics to the javaagent instrumentation
