+++ instrumentation/kafka/kafka-clients/kafka-clients-2.6/library/src/main/java/io/opentelemetry/instrumentation/kafkaclients/KafkaTelemetryBuilder.java
@@ -18,7 +18,7 @@ import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.producer.ProducerRecord;
 
 public final class KafkaTelemetryBuilder {
-  private static final String INSTRUMENTATION_NAME = "io.opentelemetry.kafka-clients-2.6";
+  static final String INSTRUMENTATION_NAME = "io.opentelemetry.kafka-clients-2.6";
 
   private final OpenTelemetry openTelemetry;
   private final List<AttributesExtractor<ProducerRecord<?, ?>, Void>> producerAttributesExtractors =
